elide and show types more carefully
The rule of thumb is, when the type is statically known from
context, elide the type. When it's not, show the type.

In practice this means usually when the a value's context
(struct field, slice element type, etc) has a static type
that's an interface, we show the value's concrete type, and
otherwise we don't.

There are a few exceptions. For instance, when a type
contains two or more adjacent levels of pointer indirection,
e.g. []**int, we show the type to be more explicit.