show common root type as first path element
If the two root values being compared, a and b, have
different types or are themselves different, output is
unchanged.

Otherwise, a and b have the same type, and any
differences are inside them, so there is a nonempty path
to each difference. We can show the root type as a
prefix of that path.

Old output: .N: 1 != 2
New output: diff_test.T.N: 1 != 2

This gives more context and makes it easier to identify
what is being diffed, especially when there is diff
output from multiple places in the code showing up in
the program's output.