fix Time transform to consider monotonic time
We have to worry about monotonic time readings, not just
location in a time.Time value. If one time has such a
reading and the other doesn't, time.Equal can return
true while the == operator returns false.

To fix this, we need to strip the monotonic time reading
(in addition to normalizing the location) before
comparing two Time values.

Note that In(UTC), which we were already doing, strips
the monotonic reading. So this change doesn't affect the
behavior. This is a documentation change to make what
we're doing more explicit.

Note also that the Round(0) here is redundant, since,
again, UTC() strips the monotonic reading. But we write
it anyway to be more expicit about the intent.