fix bug in xform logic
It was a mistake to use an empty table of xform funcs in
the equal method, which was used to check for equality
in an entire subgraph, speicifically to test whether the
transformed values are equal (and if so to return early,
ensuring we don't report any differences for the
original values).

However, this meant that equal could produce different
results than method each itself. Here's a scenario to
illustrate, based on the unit test added here:

There's a transform for type T, plus a transform for
type Time (installed by default).

* The transform for type T causes us to call equal,
  which removes all transforms for itself.
* We compare the (transformed) T values without any
  transforms.
* We walk to the Time values inside of type T.
* The Time values compare as unequal. (!)
* Method equal returns false.
* We compare the (original) T values with transforms.
* We walk to the Time values inside of type T.
* We transform the Time values and they compare as
  equal. (!)
* We report an inconsistency "(transformed values
  differ)".

The original motivation for using an empty transform
table was to prevent unbounded recursion when comparing
transformed values. (If we simply leave the same table in
place, and the transformed values are of the same type
as the original values, then the same transform func
will be used again on the transformed values, and the
process repeats.)

Instead, we add a flag to walk to skip using the
transform func for the toplevel value in the data
structure. This is a more precise change, which just
prevents the unwanted recursive call to equal and
doesn't otherwise affect the behavior. The flag is then
turned off for further recursive calls (to walk), and
the xform table is left in place, so we continue using
transforms for the rest of the data structure, and the
two methods give consistent results.

Note that there was previously an xformOk parameter, but
it was unused until now.

We also rename equal to equalAsIs to emphasize that it
doesn't apply any xform func on the toplevel value.