let xform & format funcs work on unexported fields
We need to use UnsafeAddr to get access to unexported
fields, so we can pass them to transform and format
functions.

We also need to make sure that all values are
addressable, which is necessary for UnsafeAddr. Values
obtained via addressable Values are addressable, so we
only need to worry about places where user-supplied
values enter our care: the entry points to the package
and return values from xform funcs.

Entry points are made addressable in method each, and
xform return values are made addressable in the
haveXform check, right after we get them.