sort map keys
This change also required us to make sure values we
handle are accessible (even if accessed via an
unexported field) and addressable. We can do that more
precisely by doing it at every moment it is possible to
obtain an unexported value or an unaddressable value.

For unexported values, that it every time we read an
unexported field. For unaddressable values, that is
every time we read a value out of an interface, in
addition to the entry points (where we were already
making values addressable).