remove redundant emitter type
We don't need countEmitter; the same job can more easily be
done by a sink in the normal printEmitter, which is now
renamed to simply emitter. And now that there's only one
type of emitter, we don't need the interface type any more.

The downside is more garbage building paths and emitters
inside a call to equal.